Added previous IfTile and Switch Tile functions
The previously made functions for the if tile and the s witch tile were in an older version of unity. I've added them to the current project though there name is still under the test script for the time being. Once the full Tile class is completed this will need to be chagned